# packages/backend/swagger.yaml
openapi: 3.0.0
paths:
...
  /api/posts:
    post:
      operationId: PostController_createPost
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostDto"
      responses:
        "201":
          description: Success create a post
          content:
            application/json:
              example:
                id: 1
                user:
                  id: 1
                content: Hello World
                createdAt: 2025-09-27T05:10:38.252Z
        "400":
          description: Validation failed.
        "422":
          description: Unprocessable enitity
      security:
        - bearer: []
      tags:
        - Post
...
info:
  title: NewsFeed API
  description: API documentation for NewsFeedApp
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    CreatePostDto:
      type: object
      properties:
        content:
          type: string
          example: Hello World
          description: The content of post
      required:
        - content
...